refactor: extract MessageRepository
