+++ packages/control-plane/src/session/repository.test.ts
@@ -4,13 +4,8 @@
  * Uses a mock SqlStorage to verify SQL operations are called correctly.
  */
 
-import { describe, it, expect, beforeEach, vi } from "vitest";
+import { describe, it, expect, beforeEach } from "vitest";
 import { SessionRepository } from "./repository";
-import { EventRepository } from "./event-repository";
-import {
-  AttachmentClaimConflictError,
-  SessionAttachmentRepository,
-} from "./session-attachment-repository";
 import type { SqlResult, SqlStorage } from "./sql-storage";
 
 /**
@@ -72,20 +67,10 @@ function createMockSql() {
 describe("SessionRepository", () => {
   let mock: ReturnType<typeof createMockSql>;
   let repo: SessionRepository;
-  let transactionSyncCalls: number;
 
   beforeEach(() => {
     mock = createMockSql();
-    transactionSyncCalls = 0;
-    repo = new SessionRepository(
-      mock.sql,
-      (closure) => {
-        transactionSyncCalls += 1;
-        return closure();
-      },
-      new SessionAttachmentRepository(mock.sql),
-      new EventRepository(mock.sql, (closure) => closure())
-    );
+    repo = new SessionRepository(mock.sql, (closure) => closure());
   });
 
   // === SESSION ===
@@ -340,15 +325,10 @@ describe("SessionRepository", () => {
 
     it("applies all baseline updates in one transaction", () => {
       let transactions = 0;
-      repo = new SessionRepository(
-        mock.sql,
-        (closure) => {
-          transactions += 1;
-          return closure();
-        },
-        new SessionAttachmentRepository(mock.sql),
-        new EventRepository(mock.sql, (closure) => closure())
-      );
+      repo = new SessionRepository(mock.sql, (closure) => {
+        transactions += 1;
+        return closure();
+      });
 
       repo.setSessionDiffBaselines([
         {
@@ -529,419 +509,6 @@ describe("SessionRepository", () => {
       expect(mock.calls[0].query).toContain("spawn_failure_count = COALESCE");
       expect(mock.calls[0].query).toContain("last_spawn_failure");
       expect(mock.calls[0].params).toEqual([7000]);
-    });
-  });
-
-  // === MESSAGES ===
-
-  describe("getMessageCount", () => {
-    it("returns 0 when empty", () => {
-      mock.setOne({ count: 0 });
-      expect(repo.getMessageCount()).toBe(0);
-    });
-
-    it("returns correct count", () => {
-      mock.setOne({ count: 5 });
-      expect(repo.getMessageCount()).toBe(5);
-    });
-  });
-
-  describe("getPendingOrProcessingCount", () => {
-    it("counts pending and processing messages", () => {
-      mock.setOne({ count: 3 });
-      expect(repo.getPendingOrProcessingCount()).toBe(3);
-      expect(mock.calls[0].query).toContain("'pending', 'processing'");
-    });
-  });
-
-  describe("getProcessingMessage", () => {
-    it("returns null when none processing", () => {
-      mock.setData(`SELECT id FROM messages WHERE status = 'processing' LIMIT 1`, []);
-      expect(repo.getProcessingMessage()).toBeNull();
-    });
-
-    it("returns processing message", () => {
-      mock.setData(`SELECT id FROM messages WHERE status = 'processing' LIMIT 1`, [
-        { id: "msg-1" },
-      ]);
-      expect(repo.getProcessingMessage()).toEqual({ id: "msg-1" });
-    });
-  });
-
-  describe("getNextPendingMessage", () => {
-    it("returns oldest pending message", () => {
-      const message = { id: "msg-1", created_at: 1000 };
-      // The query is dynamic, so we match by result
-      mock.setData(
-        `SELECT * FROM messages WHERE status = 'pending' ORDER BY created_at ASC, rowid ASC LIMIT 1`,
-        [message]
-      );
-      expect(repo.getNextPendingMessage()).toEqual(message);
-      expect(mock.calls[0].query).toContain("ORDER BY created_at ASC, rowid ASC");
-    });
-  });
-
-  describe("prompt queue", () => {
-    it("returns null instead of position zero for a finished idempotent message", () => {
-      mock.setData(
-        `SELECT id FROM messages WHERE status IN ('pending', 'processing')
-       ORDER BY CASE status WHEN 'processing' THEN 0 ELSE 1 END, created_at ASC, rowid ASC`,
-        [{ id: "msg-other" }]
-      );
-      expect(repo.getUnfinishedMessagePosition("msg-complete")).toBeNull();
-    });
-
-    it("projects only fields needed to render the queue", () => {
-      vi.spyOn(repo, "listUnfinishedMessages").mockReturnValue([
-        {
-          id: "msg-legacy",
-          author_id: "part-1",
-          content: "continue",
-          source: "web",
-          model: null,
-          reasoning_effort: null,
-          attachments: "{bad json",
-          callback_context: null,
-          client_request_id: null,
-          request_fingerprint: null,
-          status: "pending",
-          error_message: null,
-          stop_confirmation_deadline: null,
-          created_at: 1000,
-          started_at: null,
-          completed_at: null,
-        },
-      ]);
-
-      expect(repo.listPromptQueue()).toEqual([
-        { messageId: "msg-legacy", content: "continue", status: "pending" },
-      ]);
-    });
-  });
-
-  describe("createMessage", () => {
-    it("creates message with all fields", () => {
-      repo.createMessage({
-        id: "msg-1",
-        authorId: "p-1",
-        content: "Hello",
-        source: "web",
-        model: "claude-sonnet-4",
-        attachments: "[]",
-        callbackContext: '{"channel":"C123"}',
-        status: "pending",
-        createdAt: 1000,
-      });
-
-      expect(mock.calls.length).toBe(1);
-      expect(mock.calls[0].query).toContain("INSERT INTO messages");
-      expect(mock.calls[0].params).toEqual([
-        "msg-1",
-        "p-1",
-        "Hello",
-        "web",
-        "claude-sonnet-4",
-        null,
-        "[]",
-        '{"channel":"C123"}',
-        null,
-        null,
-        "pending",
-        1000,
-      ]);
-    });
-  });
-
-  describe("createMessageWithAttachments", () => {
-    const message = {
-      id: "msg-1",
-      authorId: "p-1",
-      content: "Look",
-      source: "web" as const,
-      status: "pending" as const,
-      createdAt: 1000,
-    };
-
-    it("claims uploads and creates the pending message in one transaction", () => {
-      let transactions = 0;
-      repo = new SessionRepository(
-        mock.sql,
-        (closure) => {
-          transactions += 1;
-          return closure();
-        },
-        new SessionAttachmentRepository(mock.sql),
-        new EventRepository(mock.sql, (closure) => closure())
-      );
-      mock.setDefaultRowsWritten(2);
-
-      repo.createMessageWithAttachments(message, ["up-1", "up-2"]);
-
-      expect(transactions).toBe(1);
-      expect(mock.calls[0].query).toContain("UPDATE attachments SET message_id");
-      expect(mock.calls[0].params).toEqual(["msg-1", "up-1", "up-2"]);
-      expect(mock.calls[1].query).toContain("INSERT INTO messages");
-      expect(mock.calls).toHaveLength(2);
-    });
-
-    it("fails before creating the message when not every upload can be claimed", () => {
-      mock.setDefaultRowsWritten(1);
-
-      expect(() => repo.createMessageWithAttachments(message, ["up-1", "up-2"])).toThrow(
-        AttachmentClaimConflictError
-      );
-      expect(mock.calls).toHaveLength(1);
-    });
-  });
-
-  describe("cancelPendingMessage", () => {
-    const statusQuery = `SELECT status, source, callback_context FROM messages WHERE id = ?`;
-
-    it("atomically releases attachments and deletes a pending message", () => {
-      mock.setData(statusQuery, [{ status: "pending", source: "web", callback_context: null }]);
-      mock.setDefaultRowsWritten(1);
-
-      expect(repo.cancelPendingMessage("msg-1")).toBe(true);
-
-      expect(transactionSyncCalls).toBe(1);
-      expect(mock.calls[1]).toMatchObject({
-        query: expect.stringContaining("UPDATE attachments SET message_id = NULL"),
-        params: ["msg-1"],
-      });
-      expect(mock.calls[2]).toMatchObject({
-        query: expect.stringContaining("DELETE FROM messages"),
-        params: ["msg-1"],
-      });
-    });
-
-    it("does not change a processing message", () => {
-      mock.setData(statusQuery, [{ status: "processing", source: "web", callback_context: null }]);
-
-      expect(repo.cancelPendingMessage("msg-1")).toBe(false);
-      expect(mock.calls).toHaveLength(1);
-    });
-
-    it("does not delete a non-web prompt that may require a callback", () => {
-      mock.setData(statusQuery, [{ status: "pending", source: "linear", callback_context: null }]);
-
-      expect(repo.cancelPendingMessage("msg-1")).toBe(false);
-      expect(mock.calls).toHaveLength(1);
-    });
-
-    it("does not delete a prompt with a completion callback", () => {
-      mock.setData(statusQuery, [
-        { status: "pending", source: "web", callback_context: '{"channel":"C1"}' },
-      ]);
-
-      expect(repo.cancelPendingMessage("msg-1")).toBe(false);
-      expect(mock.calls).toHaveLength(1);
-    });
-  });
-
-  describe("startMessageProcessing", () => {
-    it("updates processing state and materializes the user event in one transaction", () => {
-      repo.startMessageProcessing("msg-1", 2000, {
-        type: "user_message",
-        content: "Hello",
-        messageId: "msg-1",
-        timestamp: 2,
-        author: { participantId: "p-1", userId: "u-1", name: "User" },
-      });
-
-      expect(transactionSyncCalls).toBe(1);
-      expect(mock.calls[0].query).toContain("status = 'processing'");
-      expect(mock.calls[0].params).toEqual([2000, "msg-1"]);
-      expect(mock.calls[1].query).toContain("INSERT INTO events");
-      expect(mock.calls[1].params.at(-1)).toBe(2000);
-    });
-
-    it("appends the canonical event without updating legacy timeline rows", () => {
-      repo.startMessageProcessing("msg-1", 2000, {
-        type: "user_message",
-        content: "Hello",
-        messageId: "msg-1",
-        timestamp: 2,
-        author: { participantId: "p-1", userId: "u-1", name: "User" },
-      });
-
-      expect(mock.calls).toHaveLength(2);
-      expect(mock.calls[1].query).toContain("INSERT INTO events");
-      expect(mock.calls[1].query).not.toContain("UPDATE events");
-      expect(mock.calls[1].params[0]).toBe("user_message:msg-1");
-    });
-  });
-
-  describe("recordMessageCompletion", () => {
-    const messageQuery = `SELECT status, created_at, started_at FROM messages WHERE id = ?`;
-
-    it("atomically records message state and its canonical completion event", () => {
-      mock.setData(messageQuery, [{ status: "processing", created_at: 1000, started_at: 1200 }]);
-      const event = {
-        type: "execution_complete" as const,
-        messageId: "msg-1",
-        success: true,
-        sandboxId: "sb-1",
-        timestamp: 3,
-      };
-
-      expect(repo.recordMessageCompletion(event, 3000, "processing")).toEqual({
-        messageId: "msg-1",
-        messageCreatedAt: 1000,
-        messageStartedAt: 1200,
-        completedAt: 3000,
-        status: "completed",
-      });
-
-      expect(transactionSyncCalls).toBe(1);
-      expect(mock.calls[1].params).toEqual(["completed", 3000, null, "msg-1"]);
-      expect(mock.calls[2].params).toEqual([
-        "execution_complete:msg-1",
-        "execution_complete",
-        JSON.stringify(event),
-        "msg-1",
-        3000,
-      ]);
-    });
-
-    it("persists a failed outcome and error", () => {
-      mock.setData(messageQuery, [{ status: "processing", created_at: 1000, started_at: 1200 }]);
-      const event = {
-        type: "execution_complete" as const,
-        messageId: "msg-1",
-        success: false,
-        error: "Agent failed",
-        sandboxId: "sb-1",
-        timestamp: 3,
-      };
-
-      const completion = repo.recordMessageCompletion(event, 3000, "processing");
-
-      expect(completion?.status).toBe("failed");
-      expect(mock.calls[1].params).toEqual(["failed", 3000, "Agent failed", "msg-1"]);
-    });
-
-    it("does not record an outcome for a message in another state", () => {
-      mock.setData(messageQuery, [{ status: "completed", created_at: 1000, started_at: 1200 }]);
-
-      const completion = repo.recordMessageCompletion(
-        {
-          type: "execution_complete",
-          messageId: "msg-1",
-          success: true,
-          sandboxId: "sb-1",
-          timestamp: 3,
-        },
-        3000,
-        "processing"
-      );
-
-      expect(completion).toBeNull();
-      expect(mock.calls).toHaveLength(1);
-    });
-  });
-
-  describe("stop confirmation deadline", () => {
-    it("marks, reads, and clears the dedicated deadline without changing error_message", () => {
-      const query = `SELECT id, stop_confirmation_deadline FROM messages
-       WHERE stop_confirmation_deadline IS NOT NULL LIMIT 1`;
-      mock.setData(query, [{ id: "msg-1", stop_confirmation_deadline: 5000 }]);
-
-      repo.markMessageAwaitingStopConfirmation("msg-1", 5000);
-      expect(mock.calls[0].query).toContain("SET stop_confirmation_deadline = ?");
-      expect(mock.calls[0].query).not.toContain("error_message");
-      expect(repo.getMessageAwaitingStopConfirmation()).toEqual({ id: "msg-1", deadline: 5000 });
-      repo.clearMessageAwaitingStopConfirmation("msg-1");
-      expect(mock.calls[2].query).toContain("SET stop_confirmation_deadline = NULL");
-      expect(mock.calls[2].query).not.toContain("error_message");
-    });
-  });
-
-  describe("listPendingMessagesWithCreatedAt", () => {
-    it("returns pending messages in deterministic queue order", () => {
-      mock.setData(
-        `SELECT id, created_at FROM messages WHERE status = 'pending' ORDER BY created_at ASC, rowid ASC`,
-        [{ id: "msg-1", created_at: 1000 }]
-      );
-
-      expect(repo.listPendingMessagesWithCreatedAt()).toEqual([{ id: "msg-1", created_at: 1000 }]);
-
-      expect(mock.calls[0].query).toContain("SELECT id, created_at FROM messages");
-      expect(mock.calls[0].query).toContain("WHERE status = 'pending'");
-      expect(mock.calls[0].query).toContain("ORDER BY created_at ASC, rowid ASC");
-      expect(mock.calls[0].params).toEqual([]);
-    });
-  });
-
-  describe("getNextPendingMessage", () => {
-    it("uses rowid as the stable tie-breaker for equal timestamps", () => {
-      repo.getNextPendingMessage();
-
-      expect(mock.calls[0].query).toContain("ORDER BY created_at ASC, rowid ASC");
-    });
-  });
-
-  describe("listMessages", () => {
-    it("returns messages with pagination", () => {
-      repo.listMessages({ limit: 10 });
-      expect(mock.calls[0].query).toContain("ORDER BY created_at DESC");
-      expect(mock.calls[0].query).toContain("LIMIT ?");
-    });
-
-    it("filters by status when provided", () => {
-      repo.listMessages({ limit: 10, status: "pending" });
-      expect(mock.calls[0].query).toContain("status = ?");
-      expect(mock.calls[0].params).toContain("pending");
-    });
-
-    it("uses cursor for pagination", () => {
-      repo.listMessages({ limit: 10, cursor: "5000" });
-      expect(mock.calls[0].query).toContain("created_at < ?");
-      expect(mock.calls[0].params).toContain(5000);
-    });
-  });
-
-  describe("getLatestTerminalMessage", () => {
-    it("selects the newest completed or failed message", () => {
-      repo.getLatestTerminalMessage();
-
-      expect(mock.calls[0].query).toContain("status IN ('completed', 'failed')");
-      expect(mock.calls[0].query).toContain(
-        "ORDER BY COALESCE(completed_at, started_at, created_at) DESC"
-      );
-      expect(mock.calls[0].query).toContain("LIMIT 1");
-    });
-  });
-
-  // === PR HELPERS ===
-
-  describe("getProcessingMessageAuthor", () => {
-    it("returns null when no processing message", () => {
-      mock.setData(`SELECT author_id FROM messages WHERE status = 'processing' LIMIT 1`, []);
-      expect(repo.getProcessingMessageAuthor()).toBeNull();
-    });
-
-    it("returns author_id of processing message", () => {
-      mock.setData(`SELECT author_id FROM messages WHERE status = 'processing' LIMIT 1`, [
-        { author_id: "p-1" },
-      ]);
-      expect(repo.getProcessingMessageAuthor()).toEqual({ author_id: "p-1" });
-    });
-  });
-
-  describe("getMessageCallbackContext", () => {
-    it("returns null for unknown message", () => {
-      mock.setData(`SELECT callback_context, source FROM messages WHERE id = ?`, []);
-      expect(repo.getMessageCallbackContext("unknown")).toBeNull();
-    });
-
-    it("returns callback context", () => {
-      mock.setData(`SELECT callback_context, source FROM messages WHERE id = ?`, [
-        { callback_context: '{"channel":"C123"}', source: "slack" },
-      ]);
-      expect(repo.getMessageCallbackContext("msg-1")).toEqual({
-        callback_context: '{"channel":"C123"}',
-        source: "slack",
-      });
     });
   });
 });
