+++ packages/control-plane/src/session/repository.ts
@@ -6,23 +6,15 @@
  * coordinated here when they also create or update core session records.
  */
 
-import type { SessionRow, MessageRow, SandboxRow, SessionRepositoryRow } from "./types";
-import type { GitSyncStatus, SandboxEvent } from "@open-inspect/shared/types/sandbox-events";
+import type { SessionRow, SandboxRow, SessionRepositoryRow } from "./types";
+import type { GitSyncStatus } from "@open-inspect/shared/types/sandbox-events";
 import type {
   SessionStatus,
   SandboxStatus,
-  MessageStatus,
-  MessageSource,
   SpawnSource,
 } from "@open-inspect/shared/types/sessions";
-import type { PromptQueueItem } from "@open-inspect/shared/types/server-messages";
-import type { CreateEventData, EventRepository } from "./event-repository";
 import { buildSessionRepositories, type SessionRepositoryEntry } from "./repository-target";
-import type { SessionAttachmentRepository } from "./session-attachment-repository";
 import type { SqlResult, SqlStorage, TransactionSync } from "./sql-storage";
-
-type ExecutionCompleteEvent = Extract<SandboxEvent, { type: "execution_complete" }>;
-export const STOP_CONFIRMATION_TIMEOUT_MS = 15_000;
 
 /**
  * Minimal sandbox state for circuit breaker checks.
@@ -35,14 +27,6 @@ export interface SandboxCircuitBreakerState {
   snapshot_image_id: string | null;
   spawn_failure_count: number | null;
   last_spawn_failure: number | null;
-}
-
-interface RecordedMessageCompletion {
-  messageId: string;
-  messageCreatedAt: number;
-  messageStartedAt: number | null;
-  completedAt: number;
-  status: "completed" | "failed";
 }
 
 /**
@@ -95,33 +79,6 @@ export interface CreateSandboxData {
 }
 
 /**
- * Data for creating a message.
- */
-export interface CreateMessageData {
-  id: string;
-  authorId: string;
-  content: string;
-  source: MessageSource;
-  model?: string | null;
-  reasoningEffort?: string | null;
-  attachments?: string | null;
-  callbackContext?: string | null;
-  clientRequestId?: string | null;
-  requestFingerprint?: string | null;
-  status: MessageStatus;
-  createdAt: number;
-}
-
-/**
- * Options for listing messages.
- */
-export interface ListMessagesOptions {
-  cursor?: string | null;
-  limit: number;
-  status?: string | null;
-}
-
-/**
  * Data for spawn sandbox update.
  */
 export interface SpawnSandboxData {
@@ -142,12 +99,7 @@ export interface ResumeSandboxData {
 export class SessionRepository {
   constructor(
     private readonly sql: SqlStorage,
-    private readonly transactionSync: TransactionSync,
-    private readonly attachments: Pick<
-      SessionAttachmentRepository,
-      "claimForMessage" | "releaseForMessage"
-    >,
-    private readonly eventRepository: EventRepository
+    private readonly transactionSync: TransactionSync
   ) {}
 
   private rows<T>(result: SqlResult): T[] {
@@ -541,308 +493,5 @@ export class SessionRepository {
        WHERE id = (SELECT id FROM sandbox LIMIT 1)`,
       timestamp
     );
-  }
-
-  // === MESSAGES ===
-
-  getActiveDurationMs(): number {
-    const result = this.sql.exec(
-      `SELECT COALESCE(SUM(completed_at - started_at), 0) as duration_ms
-       FROM messages
-       WHERE started_at IS NOT NULL AND completed_at IS NOT NULL`
-    );
-    return (result.one() as { duration_ms: number }).duration_ms;
-  }
-
-  getMessageCount(): number {
-    const result = this.sql.exec(`SELECT COUNT(*) as count FROM messages`);
-    return (result.one() as { count: number }).count;
-  }
-
-  getPendingOrProcessingCount(): number {
-    const result = this.sql.exec(
-      `SELECT COUNT(*) as count FROM messages WHERE status IN ('pending', 'processing')`
-    );
-    return (result.one() as { count: number }).count;
-  }
-
-  getProcessingMessage(): { id: string } | null {
-    const result = this.sql.exec(`SELECT id FROM messages WHERE status = 'processing' LIMIT 1`);
-    const rows = result.toArray() as Array<{ id: string }>;
-    return rows[0] ?? null;
-  }
-
-  getMessageAwaitingStopConfirmation(): { id: string; deadline: number } | null {
-    const result = this.sql.exec(
-      `SELECT id, stop_confirmation_deadline FROM messages
-       WHERE stop_confirmation_deadline IS NOT NULL LIMIT 1`
-    );
-    const row = (result.toArray() as Array<{ id: string; stop_confirmation_deadline: number }>)[0];
-    return row ? { id: row.id, deadline: row.stop_confirmation_deadline } : null;
-  }
-
-  markMessageAwaitingStopConfirmation(messageId: string, deadline: number): void {
-    this.sql.exec(
-      `UPDATE messages SET stop_confirmation_deadline = ? WHERE id = ?`,
-      deadline,
-      messageId
-    );
-  }
-
-  clearMessageAwaitingStopConfirmation(messageId: string): void {
-    this.sql.exec(`UPDATE messages SET stop_confirmation_deadline = NULL WHERE id = ?`, messageId);
-  }
-
-  getProcessingMessageWithCreatedAt(): { id: string; created_at: number } | null {
-    const result = this.sql.exec(
-      `SELECT id, created_at FROM messages WHERE status = 'processing' LIMIT 1`
-    );
-    const rows = result.toArray() as Array<{ id: string; created_at: number }>;
-    return rows[0] ?? null;
-  }
-
-  getProcessingMessageWithStartedAt(): { id: string; started_at: number } | null {
-    const result = this.sql.exec(
-      `SELECT id, started_at FROM messages WHERE status = 'processing' LIMIT 1`
-    );
-    const rows = result.toArray() as Array<{ id: string; started_at: number }>;
-    return rows[0] ?? null;
-  }
-
-  getNextPendingMessage(): MessageRow | null {
-    const result = this.sql.exec(
-      `SELECT * FROM messages WHERE status = 'pending' ORDER BY created_at ASC, rowid ASC LIMIT 1`
-    );
-    const rows = this.rows<MessageRow>(result);
-    return rows[0] ?? null;
-  }
-
-  getMessageByClientRequestId(clientRequestId: string): MessageRow | null {
-    const result = this.sql.exec(
-      `SELECT * FROM messages WHERE client_request_id = ? LIMIT 1`,
-      clientRequestId
-    );
-    return this.rows<MessageRow>(result)[0] ?? null;
-  }
-
-  getUnfinishedMessagePosition(messageId: string): number | null {
-    const result = this.sql.exec(
-      `SELECT id FROM messages WHERE status IN ('pending', 'processing')
-       ORDER BY CASE status WHEN 'processing' THEN 0 ELSE 1 END, created_at ASC, rowid ASC`
-    );
-    const index = (result.toArray() as Array<{ id: string }>).findIndex(
-      (row) => row.id === messageId
-    );
-    return index < 0 ? null : index + 1;
-  }
-
-  listUnfinishedMessages(): MessageRow[] {
-    const result = this.sql.exec(
-      `SELECT * FROM messages WHERE status IN ('pending', 'processing')
-       ORDER BY CASE status WHEN 'processing' THEN 0 ELSE 1 END, created_at ASC, rowid ASC`
-    );
-    return this.rows<MessageRow>(result);
-  }
-
-  listPromptQueue(): PromptQueueItem[] {
-    return this.listUnfinishedMessages().map((message) => ({
-      messageId: message.id,
-      content: message.content,
-      status: message.status as "pending" | "processing",
-    }));
-  }
-
-  cancelPendingMessage(messageId: string): boolean {
-    return this.transactionSync(() => {
-      const result = this.sql.exec(
-        `SELECT status, source, callback_context FROM messages WHERE id = ?`,
-        messageId
-      );
-      const message = (
-        result.toArray() as Array<{
-          status: MessageStatus;
-          source: string;
-          callback_context: string | null;
-        }>
-      )[0];
-      if (
-        message?.status !== "pending" ||
-        message.source !== "web" ||
-        message.callback_context !== null
-      ) {
-        return false;
-      }
-
-      this.attachments.releaseForMessage(messageId);
-      const deleted = this.sql.exec(
-        `DELETE FROM messages WHERE id = ? AND status = 'pending'`,
-        messageId
-      );
-      deleted.toArray();
-      return deleted.rowsWritten === 1;
-    });
-  }
-
-  getMessageCallbackContext(
-    messageId: string
-  ): { callback_context: string | null; source: string | null } | null {
-    const result = this.sql.exec(
-      `SELECT callback_context, source FROM messages WHERE id = ?`,
-      messageId
-    );
-    const rows = result.toArray() as Array<{
-      callback_context: string | null;
-      source: string | null;
-    }>;
-    return rows[0] ?? null;
-  }
-
-  createMessage(data: CreateMessageData): void {
-    this.sql.exec(
-      `INSERT INTO messages (id, author_id, content, source, model, reasoning_effort, attachments, callback_context, client_request_id, request_fingerprint, status, created_at)
-       VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)`,
-      data.id,
-      data.authorId,
-      data.content,
-      data.source,
-      data.model ?? null,
-      data.reasoningEffort ?? null,
-      data.attachments ?? null,
-      data.callbackContext ?? null,
-      data.clientRequestId ?? null,
-      data.requestFingerprint ?? null,
-      data.status,
-      data.createdAt
-    );
-  }
-
-  /** Persist a message, its attachments, and canonical timeline event atomically. */
-  createMessageWithAttachments(
-    data: CreateMessageData,
-    attachmentIds: string[],
-    event?: CreateEventData
-  ): void {
-    this.transactionSync(() => {
-      this.attachments.claimForMessage(data.id, attachmentIds);
-      this.createMessage(data);
-      if (event) this.eventRepository.createEvent(event);
-    });
-  }
-
-  startMessageProcessing(
-    messageId: string,
-    startedAt: number,
-    userMessageEvent: Extract<SandboxEvent, { type: "user_message" }>
-  ): void {
-    this.transactionSync(() => {
-      this.sql.exec(
-        `UPDATE messages SET status = 'processing', started_at = ? WHERE id = ?`,
-        startedAt,
-        messageId
-      );
-      this.eventRepository.createEvent({
-        id: `user_message:${messageId}`,
-        type: "user_message",
-        data: JSON.stringify(userMessageEvent),
-        messageId,
-        createdAt: startedAt,
-      });
-    });
-  }
-
-  updateMessageToPending(messageId: string): void {
-    this.sql.exec(
-      `UPDATE messages SET status = 'pending', started_at = NULL WHERE id = ? AND status = 'processing'`,
-      messageId
-    );
-  }
-
-  recordMessageCompletion(
-    event: ExecutionCompleteEvent,
-    completedAt: number,
-    expectedStatus: "pending" | "processing"
-  ): RecordedMessageCompletion | null {
-    return this.transactionSync(() => {
-      const result = this.sql.exec(
-        `SELECT status, created_at, started_at FROM messages WHERE id = ?`,
-        event.messageId
-      );
-      const message = (
-        result.toArray() as Array<{
-          status: MessageStatus;
-          created_at: number;
-          started_at: number | null;
-        }>
-      )[0];
-      if (!message || message.status !== expectedStatus) return null;
-
-      const status = event.success ? "completed" : "failed";
-      this.sql.exec(
-        `UPDATE messages SET status = ?, completed_at = ?, error_message = ? WHERE id = ?`,
-        status,
-        completedAt,
-        event.success ? null : (event.error ?? null),
-        event.messageId
-      );
-      this.eventRepository.upsertExecutionCompleteEvent(event.messageId, event, completedAt);
-
-      return {
-        messageId: event.messageId,
-        messageCreatedAt: message.created_at,
-        messageStartedAt: message.started_at,
-        completedAt,
-        status,
-      };
-    });
-  }
-
-  listPendingMessagesWithCreatedAt(): Array<{ id: string; created_at: number }> {
-    const result = this.sql.exec(
-      `SELECT id, created_at FROM messages WHERE status = 'pending' ORDER BY created_at ASC, rowid ASC`
-    );
-    return result.toArray() as Array<{ id: string; created_at: number }>;
-  }
-
-  listMessages(options: ListMessagesOptions): MessageRow[] {
-    // WHERE 1=1 allows appending AND clauses unconditionally
-    let query = `SELECT * FROM messages WHERE 1=1`;
-    const params: (string | number)[] = [];
-
-    if (options.status) {
-      query += ` AND status = ?`;
-      params.push(options.status);
-    }
-
-    if (options.cursor) {
-      query += ` AND created_at < ?`;
-      params.push(parseInt(options.cursor));
-    }
-
-    query += ` ORDER BY created_at DESC LIMIT ?`;
-    params.push(options.limit + 1);
-
-    const result = this.sql.exec(query, ...params);
-    return this.rows<MessageRow>(result);
-  }
-
-  getLatestTerminalMessage(): MessageRow | null {
-    const result = this.sql.exec(
-      `SELECT * FROM messages
-       WHERE status IN ('completed', 'failed')
-       ORDER BY COALESCE(completed_at, started_at, created_at) DESC, created_at DESC, id DESC
-       LIMIT 1`
-    );
-    const rows = this.rows<MessageRow>(result);
-    return rows[0] ?? null;
-  }
-
-  // === PR HELPERS ===
-
-  getProcessingMessageAuthor(): { author_id: string } | null {
-    const result = this.sql.exec(
-      `SELECT author_id FROM messages WHERE status = 'processing' LIMIT 1`
-    );
-    const rows = result.toArray() as Array<{ author_id: string }>;
-    return rows[0] ?? null;
   }
 }
